Refactorización: Descomposicion y Redistribución del método statement()

+++ src/org/pris2/actividad1/Customer.java
@@ -57,7 +57,7 @@ public class Customer {
 		return result;
 	}
 
-	private int amountFor(Rental each) {
+	private double amountFor(Rental each) {
 		int thisAmount = 0;
 		switch (each.getMovie().getPriceCode()) {
 			case Movie.REGULAR:
